Refactor evaluators into separate parser/tokenizer classes
